+++ hasklepias-core/src/Cohort/Core.hs
@@ -8,11 +8,14 @@ Maintainer  : bsaul@novisci.com
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TupleSections #-}
+{-# LANGUAGE LambdaCase #-}
 -- {-# LANGUAGE Safe #-}
 
 module Cohort.Core
   ( Subject(..)
-  , ID
+  , SubjectID
+  , ObsID
   , Population(..)
   , ObsUnit(..)
   , CohortData(..)
@@ -20,10 +23,8 @@ module Cohort.Core
   , CohortSpec
   , CohortSetSpec
   , CohortSet(..)
-  , AttritionInfo(..)
-  , AttritionLevel(..)
+  , makeObsID
   , specifyCohort
-  , makeObsUnitFeatures
   , evalCohort
   , getCohortIDs
   , getCohortDataIDs
@@ -35,13 +36,19 @@ module Cohort.Core
   , getCohortSet
   ) where
 
+import           Cohort.Attrition
 import           Cohort.Criteria                ( CohortStatus(..)
                                                 , Criteria(getCriteria)
                                                 , checkCohortStatus
                                                 , initStatusInfo
                                                 )
-import           Cohort.Index                   ( Index(..)
-                                                , makeIndex
+import           Cohort.Index
+import           Control.Applicative            ( (<$>)
+                                                , pure
+                                                )
+import           Control.Monad                  ( (=<<)
+                                                , Functor(fmap)
+                                                , join
                                                 )
 import           Data.Aeson                     ( FromJSON
                                                 , ToJSON(..)
@@ -49,47 +56,51 @@ import           Data.Aeson                     ( FromJSON
 import           Data.Bool                      ( Bool )
 import           Data.Eq                        ( Eq )
 import           Data.Foldable                  ( Foldable(length) )
-import           Data.Function                  ( ($) )
-import           Data.Functor                   ( (<$>)
-                                                , Functor(fmap)
+import           Data.Function                  ( ($)
+                                                , (.)
+                                                , flip
                                                 )
-import           Data.List                      ( replicate
-                                                , zip
-                                                , zipWith
-                                                )
-import qualified Data.List.NonEmpty            as NEL
-                                                ( NonEmpty(..)
-                                                )
+import           Data.List                      ( zipWith )
 import           Data.Map.Strict               as Map
-                                                ( Map
-                                                , fromListWith
-                                                , unionsWith
-                                                )
+                                                ( Map )
 import           Data.Maybe                     ( Maybe(..)
-                                                , maybe
-                                                , catMaybes
+                                                , mapMaybe
                                                 )
 import           Data.Monoid                    ( mempty )
 import           Data.Ord                       ( Ord(..) )
 import           Data.Semigroup                 ( Semigroup((<>)) )
 import qualified Data.Set                      as Set
                                                 ( Set
+                                                , toList
                                                 )
-import           Data.Text                      ( Text )
-import           Data.Tuple                     ( uncurry )
+import           Data.Text                      ( Text
+                                                , pack
+                                                )
+import           Data.Traversable               ( sequenceA )
+import           Data.Tuple                     ( curry )
 import           GHC.Exts                       ( IsList(..) )
 import           GHC.Generics                   ( Generic )
-import           GHC.Int                        ( Int )
-import           GHC.Num                        ( Natural
-                                                , Num((+))
-                                                )
+import           GHC.Num                        ( Natural )
 import           GHC.Show                       ( Show(..) )
 import           Safe                           ( headMay )
+import           Test.Tasty.Ingredients.ConsoleReporter
+                                                ( Statistics(statFailures) )
+
+
 -- | A subject identifier. Currently, simply @Text@.
-type ID = Text
+type SubjectID = Text
+
+-- | An observational unit identifier. Currently, simply @Text@.
+-- type ObsID = Text
+newtype ObsID = MkObsID (SubjectID, Natural)
+  deriving (Eq, Show, Generic)
+
+-- | Smart constructor for @'ObsID'@.
+makeObsID :: Natural -> SubjectID -> ObsID
+makeObsID = (flip . curry) MkObsID
 
 -- | A subject is just a pair of @ID@ and data.
-newtype Subject d = MkSubject (ID, d)
+newtype Subject d = MkSubject (SubjectID, d)
     deriving (Eq, Show, Generic)
 
 instance Functor Subject where
@@ -98,7 +109,7 @@ instance Functor Subject where
 instance (FromJSON d) => FromJSON (Subject d) where
 
 -- | A population is a list of @'Subject'@s
-newtype Population d = MkPopulation  [Subject d]
+newtype Population d = MkPopulation [Subject d]
     deriving (Eq, Show, Generic)
 
 instance Functor Population where
@@ -106,12 +117,16 @@ instance Functor Population where
 
 instance (FromJSON d) => FromJSON (Population d) where
 
--- | An observational unit is what a subject may be transformed into.
+-- | A @'Subject'@ may be mapped to zero of more observational units (one per
+--   @'Index'@ in the @'IndexSet'@).
 data ObsUnit d = MkObsUnit
-  { obsID   :: ID
+  { obsID   :: ObsID
   , obsData :: d
   }
   deriving (Eq, Show, Generic)
+
+instance Functor ObsUnit where
+  fmap f (MkObsUnit id x) = MkObsUnit id (f x)
 
 -- | A container for CohortData
 newtype CohortData d = MkCohortData { getObsData :: [ObsUnit d] }
@@ -126,132 +141,12 @@ newtype Cohort d = MkCohort (AttritionInfo, CohortData d)
 getAttritionInfo :: Cohort d -> AttritionInfo
 getAttritionInfo (MkCohort (x, _)) = x
 
--- | Unpacks a @'Population'@ to a list of subjects.
-getPopulation :: Population d -> [Subject d]
-getPopulation (MkPopulation x) = x
-
--- | Gets the data out of  a @'Subject'@.
-getSubjectData :: Subject d -> d
-getSubjectData (MkSubject (_, x)) = x
-
--- | Tranforms a @'Subject'@ into a @'ObsUnit'@.
-makeObsUnitFeatures :: (d1 -> d0) -> Subject d1 -> ObsUnit d0
-makeObsUnitFeatures f (MkSubject (id, dat)) = MkObsUnit id (f dat)
-
--- | A cohort specification consist of two functions: one that transforms a subject's
--- input data into a @'Criteria'@ and another that transforms a subject's input data
--- into the desired return type.
-data CohortSpec d1 d0 = MkCohortSpec
-  { runCriteria :: d1 -> Criteria
-        -- (Feature (Index i a))
-  , runFeatures :: d1 -> d0
-  }
-
--- | Creates a @'CohortSpec'@.
-specifyCohort :: (d1 -> Criteria) -> (d1 -> d0) -> CohortSpec d1 d0
-specifyCohort = MkCohortSpec
-
--- | Evaluates the @'runCriteria'@ of a @'CohortSpec'@ on a @'Population'@ to 
--- return a list of @Subject Criteria@ (one per subject in the population). 
-evalCriteria :: CohortSpec d1 d0 -> Population d1 -> [Subject Criteria]
-evalCriteria (MkCohortSpec runCrit _) (MkPopulation pop) =
-  fmap (fmap runCrit) pop
-
--- | Convert a list of @Subject Criteria@ into a list of @Subject CohortStatus@
-evalCohortStatus :: [Subject Criteria] -> [Subject CohortStatus]
-evalCohortStatus = fmap (fmap checkCohortStatus)
-
--- | Runs the input function which transforms a subject into an observational unit. 
--- If the subeject is excluded, the result is @Nothing@; otherwise it is @Just@ 
--- an observational unit.
-evalSubjectCohort
-  :: (d1 -> d0) -> Subject CohortStatus -> Subject d1 -> Maybe (ObsUnit d0)
-evalSubjectCohort f (MkSubject (id, status)) subjData = case status of
-  Included     -> Just $ makeObsUnitFeatures f subjData
-  ExcludedBy _ -> Nothing
-
--- | A type which collects counts of a 'CohortStatus'
-data AttritionLevel = MkAttritionLevel
-  { attritionLevel :: CohortStatus
-  , attritionCount :: Natural
-  }
-  deriving (Eq, Show, Generic)
-
--- | Ordering of @AttritionLevel@ is based on the value of its 'attritionLevel'. 
-instance Ord AttritionLevel where
-  compare (MkAttritionLevel l1 _) (MkAttritionLevel l2 _) = compare l1 l2
-
--- | NOTE: the @Semigroup@ instance prefers the 'attritionLevel' from the left,
---   so be sure that you're combining 
-instance Semigroup AttritionLevel where
-  (<>) (MkAttritionLevel l1 c1) (MkAttritionLevel _ c2) =
-    MkAttritionLevel l1 (c1 + c2)
-
--- | A type which collects the counts of subjects included or excluded.
-data AttritionInfo = MkAttritionInfo
-  { totalProcessed :: Int
-  , attritionInfo  :: Set.Set AttritionLevel
-  }
-  deriving (Eq, Show, Generic)
-
-setAttrLevlToMap :: Set.Set AttritionLevel -> Map.Map CohortStatus Natural
-setAttrLevlToMap x =
-  fromList $ (\(MkAttritionLevel l c) -> (l, c)) <$> toList x
-
-mapToSetAttrLevel :: Map.Map CohortStatus Natural -> Set.Set AttritionLevel
-mapToSetAttrLevel x = fromList $ uncurry MkAttritionLevel <$> toList x
-
--- | Two @AttritionInfo@ values can be combined, but this meant for combining
---   attrition info from the same set of @Criteria@.
-instance Semigroup AttritionInfo where
-  (<>) (MkAttritionInfo t1 i1) (MkAttritionInfo t2 i2) = MkAttritionInfo
-    (t1 + t2)
-    ( mapToSetAttrLevel
-    $ unionsWith (+) [setAttrLevlToMap i1, setAttrLevlToMap i2]
-    )
-
--- Initializes @AttritionInfo@ from a @'Criteria'@.
-initAttritionInfo :: Criteria -> Map.Map CohortStatus Natural
-initAttritionInfo x = fromList
-  $ zip (toList (initStatusInfo x)) (replicate (length (getCriteria x)) 0)
-
--- An internal function used to measure attrition for a cohort.
-measureAttrition
-  :: Maybe Criteria -> [Subject CohortStatus] -> AttritionInfo
-measureAttrition c l =
-   MkAttritionInfo (length l) $ mapToSetAttrLevel $ unionsWith
-    (+)
-    [ maybe mempty initAttritionInfo c
-    , Map.fromListWith (+) $ fmap (\x -> (getSubjectData x, 1)) l
-    , fromList [(Included, 0)]
-        -- including Included in the case that none of the evaluated criteria
-        -- have status Include
-    ]
-
--- | The internal function to evaluate a @'CohortSpec'@ on a @'Population'@. 
-evalUnits
-  :: CohortSpec d1 d0 -> Population d1 -> (AttritionInfo, CohortData d0)
-evalUnits spec pop =
-  ( measureAttrition fcrit statuses
-  , MkCohortData $ catMaybes $ zipWith (evalSubjectCohort (runFeatures spec))
-                                       statuses
-                                       (getPopulation pop)
-  )
- where
-  crits    = evalCriteria spec pop
-  fcrit    = fmap getSubjectData (headMay crits)
-  statuses = evalCohortStatus crits
-
--- | Evaluates a @'CohortSpec'@ on a @'Population'@.
-evalCohort :: CohortSpec d1 d0 -> Population d1 -> Cohort d0
-evalCohort s p = MkCohort $ evalUnits s p
-
 -- | Get IDs from 'CohortData'.
-getCohortDataIDs :: CohortData d -> [ID]
+getCohortDataIDs :: CohortData d -> [ObsID]
 getCohortDataIDs (MkCohortData x) = fmap obsID x
 
 -- | Get IDs from a cohort.
-getCohortIDs :: Cohort d -> [ID]
+getCohortIDs :: Cohort d -> [ObsID]
 getCohortIDs (MkCohort (_, dat)) = getCohortDataIDs dat
 
 -- | Get data from a cohort.
@@ -261,6 +156,102 @@ getCohortDataData (MkCohortData x) = fmap obsData x
 -- | Get data from a cohort.
 getCohortData :: Cohort d -> [d]
 getCohortData (MkCohort (_, dat)) = getCohortDataData dat
+
+-- | Unpacks a @'Population'@ to a list of subjects.
+getPopulation :: Population d -> [Subject d]
+getPopulation (MkPopulation x) = x
+
+-- | Gets the data out of  a @'Subject'@.
+getSubjectData :: Subject d -> d
+getSubjectData (MkSubject (_, x)) = x
+
+-- | A cohort specification consist of two functions: one that transforms a subject's
+-- input data into a @'Criteria'@ and another that transforms a subject's input data
+-- into the desired return type.
+data CohortSpec d1 d0 i a = MkCohortSpec
+  { runIndices  :: d1 -> IndexSet i a
+  , runCriteria :: d1 -> Criteria
+  , runFeatures :: Index i a -> d1 -> d0
+  }
+
+-- | Creates a @'CohortSpec'@.
+specifyCohort
+  :: (d1 -> IndexSet i a)
+  -> (d1 -> Criteria)
+  -> (Index i a -> d1 -> d0)
+  -> CohortSpec d1 d0 i a
+specifyCohort = MkCohortSpec
+
+-- internal functions to evalIndices
+evalIndicesOnSubjects
+  :: CohortSpec d1 d0 i a -> Population d1 -> [(IndexSet i a, Subject d1)]
+evalIndicesOnSubjects spec (MkPopulation pop) =
+  fmap (\x -> (runIndices spec (getSubjectData x), x)) pop
+
+data Step1 d i a = MkStep1 (ObsUnit d) (Maybe (Index i a))
+
+doStep1 :: (IndexSet i a, Subject d1) -> [Step1 d1 i a]
+doStep1 (MkIndexSet indices, MkSubject (id, d)) = case indices of
+  Nothing  -> pure $ MkStep1 (MkObsUnit (makeObsID 1 id) d) Nothing
+  Just set -> zipWith
+    (\x y -> MkStep1 (MkObsUnit (makeObsID y id) d) (Just x))
+    (Set.toList set)
+    [1 ..]
+
+-- | Evaluates the @'runIndices'@ of a @'CohortSpec'@ on a @'Population'@ to 
+-- return a list of @(Index i a, ObsUnit d)@ (one per observational unit). 
+evalIndices :: CohortSpec d1 d0 i a -> Population d1 -> [Step1 d1 i a]
+evalIndices spec x = doStep1 =<< evalIndicesOnSubjects spec x
+
+newtype Step2 d i a = MkStep2 (Criteria, Step1 d i a)
+
+-- | Evaluates the @'runCriteria'@ of a @'CohortSpec'@ on a @'Population'@ to 
+-- return a list of @(Criteria, Index i a, ObsUnit d1)@. 
+evalCriteria :: CohortSpec d1 d0 i a -> [Step1 d1 i a] -> [Step2 d1 i a]
+evalCriteria spec = fmap
+  (\(MkStep1 ou mi) -> MkStep2 (runCriteria spec (obsData ou), MkStep1 ou mi))
+
+data Step3 d i a =
+    I1 CohortStatus
+  | I2 CohortStatus (ObsUnit d) (Index i a)
+
+doStep3 :: Step2 d i a -> Step3 d i a
+doStep3 (MkStep2 (c, MkStep1 ou i)) = case (status, i) of
+  (Included, Just index) -> I2 status ou index
+  _                      -> I1 status
+  where status = checkCohortStatus i c
+
+step3toStatus :: Step3 d i a -> CohortStatus
+step3toStatus (I1 x    ) = x
+step3toStatus (I2 x _ _) = x
+
+-- | Convert a list of @(Criteria, Index i a, ObsUnit d)@ into a list of 
+--   @(CohortStatus, Index i a, ObsUnit d)@
+evalCohortStatus :: [Step2 d i a] -> [Step3 d i a]
+evalCohortStatus = fmap doStep3
+
+-- | Internal function for mapping an @ObsUnit d1@ along with its corresponding 
+--   @CohortStatus@ and @Index@ into @Maybe (ObsUnit d0)@. 
+evalFeatures
+  :: (Index i a -> d1 -> d0)
+  -> Step3 d1 i a
+  -> Maybe (ObsUnit d0)
+evalFeatures f x = case x of
+  I2 status obsUnit index -> Just $ fmap (f index) obsUnit
+  _                       -> Nothing
+
+-- | The internal function to evaluate a @'CohortSpec'@ on a @'Population'@. 
+evalCohort :: CohortSpec d1 d0 i a -> Population d1 -> Cohort d0
+evalCohort spec pop = MkCohort
+  ( measureAttrition firstCrit statuses
+  , MkCohortData $ mapMaybe (evalFeatures (runFeatures spec)) unitsWithStatus
+  )
+ where
+  indices         = evalIndices spec pop
+  unitsWithCrits  = evalCriteria spec indices
+  unitsWithStatus = evalCohortStatus unitsWithCrits
+  firstCrit       = (\(MkStep2 (c, _)) -> c) <$> headMay unitsWithCrits
+  statuses        = step3toStatus <$> unitsWithStatus
 
 {-| A container hold multiple cohorts of the same type. The key is the name of 
     the cohort; value is a cohort.
@@ -272,15 +263,16 @@ newtype CohortSet d = MkCohortSet (Map Text (Cohort d))
 getCohortSet :: CohortSet d -> Map Text (Cohort d)
 getCohortSet (MkCohortSet x) = x
 
-{-| Key/value pairs of 'CohortSpec's. The keys are the names of the cohorts.
--}
-newtype CohortSetSpec i d = MkCohortSetSpec (Map Text (CohortSpec i d))
+-- | Key/value pairs of 'CohortSpec's. The keys are the names of the cohorts.
+newtype CohortSetSpec d1 d0 i a = MkCohortSetSpec (Map Text (CohortSpec d1 d0 i a))
 
 -- | Make a set of 'CohortSpec's from list input.
-makeCohortSpecs :: [(Text, d1 -> Criteria, d1 -> d0)] -> CohortSetSpec d1 d0
-makeCohortSpecs l =
-  MkCohortSetSpec $ fromList (fmap (\(n, c, f) -> (n, specifyCohort c f)) l)
+makeCohortSpecs
+  :: [(Text, d1 -> IndexSet i a, d1 -> Criteria, Index i a -> d1 -> d0)]
+  -> CohortSetSpec d1 d0 i a
+makeCohortSpecs l = MkCohortSetSpec
+  $ fromList (fmap (\(n, i, c, f) -> (n, specifyCohort i c f)) l)
 
 -- | Evaluates a @'CohortSetSpec'@ on a @'Population'@.
-evalCohortSet :: CohortSetSpec d1 d0 -> Population d1 -> CohortSet d0
+evalCohortSet :: CohortSetSpec d1 d0 i a -> Population d1 -> CohortSet d0
 evalCohortSet (MkCohortSetSpec s) p = MkCohortSet $ fmap (`evalCohort` p) s